+++ torch/csrc/jit/passes/frozen_conv_add_relu_fusion.h
@@ -6,6 +6,8 @@
 namespace torch {
 namespace jit {
 
+extern std::function<void(std::shared_ptr<Graph>&)> _fuseFrozenConvAddReluImpl;
+
 TORCH_API void FuseFrozenConvAddRelu(std::shared_ptr<Graph>& graph);
 
 } // namespace jit
